+++ ui/mods/com.pa.quitch.gwaioverhaul/difficulty.js
@@ -53,7 +53,8 @@ requireGW([
         econBase: 0.5,
         econRatePerDist: 0.05,
         max_basic_fabbers: 10,
-        max_advanced_fabbers: 10
+        max_advanced_fabbers: 10,
+        ffa_chance: 5
       },
       {
         // Bronze
@@ -72,7 +73,8 @@ requireGW([
         econBase: 0.7,
         econRatePerDist: 0.05,
         max_basic_fabbers: 20,
-        max_advanced_fabbers: 20
+        max_advanced_fabbers: 20,
+        ffa_chance: 10
       },
       {
         // Silver
@@ -95,7 +97,8 @@ requireGW([
         econBase: 0.8,
         econRatePerDist: 0.1,
         max_basic_fabbers: 30,
-        max_advanced_fabbers: 30
+        max_advanced_fabbers: 30,
+        ffa_chance: 10
       },
       {
         // Gold
@@ -118,7 +121,8 @@ requireGW([
         econBase: 0.8,
         econRatePerDist: 0.1,
         max_basic_fabbers: 40,
-        max_advanced_fabbers: 40
+        max_advanced_fabbers: 40,
+        ffa_chance: 10
       },
       {
         // Platinum
@@ -141,7 +145,8 @@ requireGW([
         econBase: 0.8,
         econRatePerDist: 0.2,
         max_basic_fabbers: 50,
-        max_advanced_fabbers: 50
+        max_advanced_fabbers: 50,
+        ffa_chance: 10
       },
       {
         // Uber
@@ -164,7 +169,8 @@ requireGW([
         econBase: 10,
         econRatePerDist: 0,
         max_basic_fabbers: 100,
-        max_advanced_fabbers: 100
+        max_advanced_fabbers: 100,
+        ffa_chance: 10
       }
     ]
 
@@ -173,7 +179,9 @@ requireGW([
     Replace GW.balance.difficultyInfo with difficultyInfo
     Limit data passed to SetAIData
     Minions use minion colour not worker colour
+    Add Foes which represent a second faction contesting the system
     */
+
     var baseNeutralStars = 2;
 
     model.makeGame = function () {
@@ -335,6 +343,12 @@ requireGW([
           //console.log("AI DIFF END: ");
         };
 
+        var aiFactions = _.range(GWFactions.length);
+        aiFactions.splice(model.playerFactionIndex(), 1);
+        //console.log("AI factions excluding player faction", aiFactions);
+        var sizeMod = GW.balance.galaxySizeDiffMod[model.newGameSizeIndex() || 0];
+        //console.log("FFA chance", diffInfo.ffa_chance * sizeMod);
+
         _.forEach(teamInfo, function (info) {
           if (info.boss) {
             setAIData(info.boss, maxDist, true);
@@ -358,6 +372,17 @@ requireGW([
                 worker.ai.minions.push(mnn);
               });
             }
+            // Assign foes
+            //console.log("AI faction", worker.ai.faction);
+            if (Math.random() * 100 <= diffInfo.ffa_chance * sizeMod) {
+              var hostileFactions = _.sample(_.without(aiFactions, worker.ai.faction));
+              //console.log("Chosen foe", hostileFactions);
+              worker.ai.foes = [];
+              var fnn = _.sample(GWFactions[hostileFactions].minions);
+              setAIData(fnn, dist, false);
+              fnn.color = fnn.color || worker.ai.color;
+              worker.ai.foes.push(fnn);
+            };
           });
         });